# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: AI Events Organizer API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenBody"
      responses:
        "200":
          description: New tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /ai/generate-description:
    post:
      operationId: generateEventDescription
      tags: [ai]
      summary: AI-generate an event description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateDescriptionBody"
      responses:
        "200":
          description: Generated description
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedDescription"
  /ai/suggest-schedule:
    post:
      operationId: suggestEventSchedule
      tags: [ai]
      summary: AI-suggest optimal event schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuggestScheduleBody"
      responses:
        "200":
          description: Scheduling suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleSuggestion"
  /ai/estimate-budget:
    post:
      operationId: estimateBudget
      tags: [ai]
      summary: AI-estimate event budget breakdown
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateBudgetBody"
      responses:
        "200":
          description: Budget estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetEstimate"
  /ai/suggest-themes:
    post:
      operationId: suggestThemes
      tags: [ai]
      summary: AI-suggest event themes and improvements
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuggestThemesBody"
      responses:
        "200":
          description: Theme suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThemeSuggestions"
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: AI event planning chat assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── AUTH ──────────────────────────────────
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [admin, organizer, attendee]
      required:
        - name
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    RefreshTokenBody:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - accessToken
        - refreshToken
        - user

    # ─── USERS ────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, organizer, attendee]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
...
    GenerateDescriptionBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        location:
          type: string
        additionalContext:
          type: ["string", "null"]
      required:
        - title
        - category
        - location
    GeneratedDescription:
      type: object
      properties:
        description:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - description
        - tags
    SuggestScheduleBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        durationHours:
          type: number
        preferredDayOfWeek:
          type: ["string", "null"]
      required:
        - title
        - category
        - durationHours
    ScheduleSuggestion:
      type: object
      properties:
        suggestions:
          type: array
          items:
            type: object
            properties:
              startDate:
                type: string
                format: date-time
              endDate:
                type: string
                format: date-time
              reasoning:
                type: string
            required:
              - startDate
              - endDate
              - reasoning
      required:
        - suggestions
    EstimateBudgetBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        location:
          type: string
        attendeeCount:
          type: integer
        durationHours:
          type: number
      required:
        - title
        - category
        - location
        - attendeeCount
        - durationHours
    BudgetEstimate:
      type: object
      properties:
        totalEstimate:
          type: number
        breakdown:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              amount:
                type: number
              percentage:
                type: number
              notes:
                type: string
            required:
              - category
              - amount
              - percentage
              - notes
        currency:
          type: string
        confidence:
          type: string
      required:
        - totalEstimate
        - breakdown
        - currency
        - confidence
    SuggestThemesBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        description:
          type: string
      required:
        - title
        - category
        - description
    ThemeSuggestions:
      type: object
      properties:
        themes:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              description:
                type: string
              improvements:
                type: array
                items:
                  type: string
              engagementTips:
                type: array
                items:
                  type: string
            required:
              - name
              - description
              - improvements
              - engagementTips
      required:
        - themes
    AiChatBody:
      type: object
      properties:
        message:
          type: string
        eventContext:
          type: ["string", "null"]
        history:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
                enum: [user, assistant]
              content:
                type: string
            required:
              - role
              - content
      required:
        - message
    AiChatResponse:
      type: object
      properties:
        reply:
          type: string
        suggestions:
          type: array
          items:
            type: string
      required:
        - reply
        - suggestions